simplify call method even more; only write action plan data if values are present

+++ src/euphorie/client/browser/risk.py
@@ -355,18 +355,9 @@ class IdentificationView(RiskBase):
 
     def __call__(self):
         super(IdentificationView, self).__call__()
-        # Render the page only if the user has edit rights,
-        # otherwise redirect to the start page of the session.
-        if not self.webhelpers.can_edit_session:
-            return self.request.response.redirect(
-                "{session_url}/@@start".format(
-                    session_url=self.webhelpers.traversed_session.absolute_url()
-                )
-            )
-        if self.webhelpers.redirectOnSurveyUpdate():
-            return
+        self.check_render_condition()
+
         utils.setLanguage(self.request, self.survey, self.survey.language)
-        context = aq_inner(self.context)
         self.set_parameter_values()
 
         if self.request.method == "POST":
@@ -385,11 +376,15 @@ class IdentificationView(RiskBase):
             session = Session()
             changed = self.set_measure_data(reply, session)
 
-            new_plans, changes = self.extract_plans_from_request()
-            for plan in context.standard_measures + context.custom_measures:
-                session.delete(plan)
-            context.action_plans.extend(new_plans)
-            changed = changes or changed
+            if reply.get("answer", None):
+                # If answer is not present in the request, do not attempt to set
+                # any action-related data, since the request might have come
+                # from a sub-form.
+                new_plans, changes = self.extract_plans_from_request()
+                for plan in self.context.standard_measures + self.context.custom_measures:
+                    session.delete(plan)
+                self.context.action_plans.extend(new_plans)
+                changed = changes or changed
 
             if self.webhelpers.use_training_module:
                 if reply.get("handle_training_notes"):
@@ -476,6 +471,18 @@ class IdentificationView(RiskBase):
         if isinstance(_next, list):
             _next = _next.pop()
         return _next
+
+    def check_render_condition(self):
+        # Render the page only if the user has edit rights,
+        # otherwise redirect to the start page of the session.
+        if not self.webhelpers.can_edit_session:
+            return self.request.response.redirect(
+                "{session_url}/@@start".format(
+                    session_url=self.webhelpers.traversed_session.absolute_url()
+                )
+            )
+        if self.webhelpers.redirectOnSurveyUpdate():
+            return
 
     def set_parameter_values(self):
         appconfig = getUtility(IAppConfig)
